Replaced ListMoves delegate with an array of piece type singletons (and simplified code))
The way the delegate array and constants assigned at the top of the class (and the fact that those constants were not accessible from any other class) really bothered me. This information really doesn't belong to the board, it belongs to the program as a whole. Thus, I made a new public static class which contains this information.
I also replaced some for loops which foreach loops since they are easier to read.