Moved MakeMove()'s special cases into piece specific class methods
board.MakeMove() is no longer aware of special cases. It instead calls the function of the piece being moved and allows that piece to handle any special functionality (such as moving the rook when castling, capturing the pawn en passant, and pawn promotion) before making the desired move.
A better way to do this will probably be to simply make each piece handle the entire process. This would allow for chess960 castling (which is something I plan on implementing)